+++ CHANGELOG.md
@@ -18,11 +18,19 @@ Types of changes:
 - [It should be possible to exclude (temporary) database jobs while generating DDL.](https://github.com/paulissoft/oracle-tools/issues/22)
 - [Add SQL Datamodeler custom library and transformations scripts.](https://github.com/paulissoft/oracle-tools/issues/46)
 
-## [1.5.0] - 2022-08-03
+## [1.5.0] - 2022-08-22
 
 ### Added
 
 - [Set up Continuous Integration / Delivery / Deployment.](https://github.com/paulissoft/oracle-tools/issues/27)
+
+### Changed
+
+- [Upgrade Flyway version.](https://github.com/paulissoft/oracle-tools/issues/45)
+
+### Fixed
+
+- [When DDL is generated with the 'sort objects by dependencies' flag, an error is raised for unknown dependencies.](https://github.com/paulissoft/oracle-tools/issues/47)
 
 ## [1.4.0] - 2022-07-17